stack_alloc: relax alignment test for arrays
When we write in a cell of an array, we check that the array is aligned.
We used to check the alignment of the cell.

This allows to compile programs that write in arrays at unaligned offsets.
This is not a problem in practice, since the generated assembly
instruction accepts unaligned addresses. However, such programs have no
meaning according to the current Jasmin semantics.