Removes max_imm from the description of asm instr
This was imprecise and redundant with the description of arguments. This
allows to remove a hack from asmgen (dealing with certain kinds of MOV).

This commit also improves the error messages produced by some checks of
asmgen: it prints the list of allowed arguments to help the user
understand what the problem is.