Hacky fix to opam problem in CI
We patch the script used by opam to sandbox its subcommands
(build/install) to make $OPAMROOT available in the sandbox. It was not
the case before because of the unusual path of the opam installation (in
/builds). The proper fix is to update to opam.2.2 when it is released
and revert this commit.

(cherry picked from commit 12628a78c3a3be5f1cd53fe61f901bcb8ba183aa)