opam: mark easy-format-1.3.3 as conflicting
In CI (i.e., when dependencies are managed by nix), installation of
easy-format at version 1.3.3 fails. This OCaml library is a dependency
of `yojson` that is used in Jasmin compiler.

The cause of this failure is not really clear (might be a bug in the
opam description of the easy-format library), but this is a quick
workaround to fix the CI jobs.