Remove field id_wsize from instruction descriptors
This field is generally meaningless. It is heuristically used in
lowering on x86 to check whether an immediate value fits within the
instruction or should go through an intermediate register.

I believe that the current design is bad and that the cleaning done in
this commit is a necessary step towards a better design (that is yet to
be discovered).