Distinguish internal from external labels
External labels correspond to return addresses, or more generally to
targets of computed jumps.

External labels may be encoded as a pointer and then stored in registers
or in memory.

As the (unspecified) encoding of labels only depends of the set of
external labels, adding fresh internal labels is guaranteed to preserve
the encoding.