Fix semantics of SHLD_16 and SHRD_16
When the (masked) shift amount is larger than the operand size, the
result is undefined. This is soundly approximated in the semantics by
throwing a type error.

Note that it is not necessary to fix the EasyCrypt model: it is meant to
be correct for safe executions only.

(cherry picked from commit ea2ea7d14e7785346aa599576313e7e552998c63)