Introduce proper location info to dummy variables in x86 lowering.
Dummy variables like the OF, ZF, CF, SF flags currently have dummy variable info (which includes their location). This results in the CT checker printing dummy locations for these flags (which does not help the developer much).

This commit fixes that and passes the instruction location to the variable info (at least on x86, on ARM this is left as a TODO).