Regalloc: print liveness information
This introduces the -pliveness command-line flag to turn on printing of
liveness information during register-allocation. It prints what local
variables are live at any program point and at the end of each function
what variables of the callers are live at call-sites (hence during the
hole execution of the function).

Co-authored-by: Benjamin Grēgoire <Benjamin.Gregoire@inria.fr>