SCT-checker: accept “sct” annotation for functions
The value of this annotation is a security signature for the function.
Said signature gives the security type  of each argument and returned
value.

A security level is made of two components: one for normal, correctly
speculated executions (this component is labelled with “normal” or “n”);
the other component for mis-speculated executions (it is labelled with
“speculative” or “s”). Each component is one of:

 - “public”
 - “secret”
 - an arbitrary identifier

There are a few notational shortcuts:

 - “transient” denotes the level { normal: public, speculative: secret }
 - a lone component x denotes the pair { n: x, s: x }

A type for non-ptr variables is just a level or the special name “msf”.
A type for a ptr variable is made of two levels: one for the value
(i.e., the array), the other one for the reference (aka the pointer)
which may become transient when unspilled from memory.

The msf type qualifies values that are public in all executions (i.e.,
{n: public, s: public}) and correctly reflect the mis-speculation state
of the micro-architecture.