stack alloc: relax the checker
Before this patch, an assignment a = b, where a and b are two arrays,
was allowed only if b was fully valid, and a was marked as fully valid
too. After this patch, b does not need to be fully valid, and a is
marked as having the same validity as b.