Call pass LowerAddressing also in OCaml
The passes between stack_alloc and reg_alloc are called twice, first in
OCaml and then in Coq. Pass LowerAddressing was inserted between
stack_alloc and reg_alloc, but was only called in Coq, creating a
discrepancy between the two calls (in OCaml and in Coq). The pass is now
called also in OCaml.

Due to the way complex Coq types are extracted in OCaml, we had to make
some types simpler on the Coq side to be able to call the pass in OCaml.