Checksafety: normalize often the list of if-conditions
In the safety pre-analysis, a set of expressions is represented by a
list (without duplicates). Duplicates are tolerated during the
construction of the list which is normalized at the end of the
pre-analysis.

This runs normalization more often so as to prevent the size of this
list to blow-up.

(cherry picked from commit 7d18625bd2f4e0ee7f4bb31089fa4cffdeb66774)