Split stype into 3 types
- atype (for abstract type): syntactic types (e.g. the types of
  variables)
- ctype (for concrete type): semantical types (e.g. the types of values)
- ltype (for low type): type of assembly instructions

This allows to make atype evolve without affecting the two other types.
This prepares in particular the apparition of length variables in the
types of arrays.

atype and ctype have already diverged. In atype, the case of arrays now
stores both the cell size and the length, instead of just the number of
bytes. A predicate "convertible" is introduced to compare atypes, since
== might be too strict.

I tried to use convertible as much as possible when comparing atypes.
But var is still declared as an eqType, and == is used in some places,
in particular in the definition of eq_expr and in the code converting
between vars and regs in asm_gen.v and arch_extra.v. In the future, we
may want to change that if it causes some breakage.