Put jasmin.sty in _static
It is still moved to and referenced from
  _downloads/unique hash/jasmin.sty
by Sphinx, but the url _static/jasmin.sty is also valid. The latter is
the one to be used to reference the file from other websites (because it
does not change).

(cherry picked from commit 93f3279b29917f0b66b8f5f1f02f0e4b94761483)