Add a `#declassify` pseudo-operator.
This operator is meant to replace all uses of the “declassify”
annotation, which is now deprecated. It should be more precise and more
flexible to use. It expresses the declassification of the value it
receives as argument.

The compilation either preserves this operator or replaces it with an
other pseudo-operator, namely `#declassify_mem`. The final compilation
step (assembly generation) currently erases these operators.

The declassify-mem operator expresses the declassification of the memory
region starting at the given pointer and whose length (in bytes) is a
parameter of the operator. It is used internally only and introduced
during the compilation (stack-alloc) of declassification of arrays.

(cherry picked from commit 2a2bbf5ebeb0a47029c2387adb3f5ad3d42088c9)